+++ NEWS.md
@@ -4,6 +4,7 @@ Version 0.12.5 For Dev-C++ 7 Beta
  - fix: can't reformat code
  - enhancement: add option for setting astyle path
  - fix: wrong file wildcard (*.*) in linux
+ - fix: open terminal in linux
 
 Version 0.12.4 For Dev-C++ 7 Beta
  - change: add copyright infos to each source file